review: keep dtype=dtype, finish the playwright removal
transformers is pinned at 4.57.1, where `torch_dtype` is the deprecated
spelling (modeling_utils warns "`torch_dtype` is deprecated! Use `dtype`
instead!"). serve/api.py's existing `dtype=dtype` is the current form, so
that hunk is reverted.

The --backend change was right — render_url() only dispatches "cdp" and
raises on anything else — but three docstrings and a CLI example still
advertised playwright. Removed those too.

+++ serve/src/pixelrag_serve/api.py
@@ -704,7 +704,7 @@ def load(args):
     model = Qwen3VLForConditionalGeneration.from_pretrained(
         args.model,
         trust_remote_code=True,
-        torch_dtype=dtype,
+        dtype=dtype,
     )
     adapter_path = getattr(args, "peft_adapter", None)
     if adapter_path:
